+++ src/Aws/DataPipeline/Resources/datapipeline-2012-10-29.php
@@ -29,6 +29,26 @@ return array (
             'https' => true,
             'hostname' => 'datapipeline.us-east-1.amazonaws.com',
         ),
+        'us-west-2' => array(
+            'http' => false,
+            'https' => true,
+            'hostname' => 'datapipeline.us-west-2.amazonaws.com',
+        ),
+        'eu-west-1' => array(
+            'http' => false,
+            'https' => true,
+            'hostname' => 'datapipeline.eu-west-1.amazonaws.com',
+        ),
+        'ap-southeast-2' => array(
+            'http' => false,
+            'https' => true,
+            'hostname' => 'datapipeline.ap-southeast-2.amazonaws.com',
+        ),
+        'ap-northeast-1' => array(
+            'http' => false,
+            'https' => true,
+            'hostname' => 'datapipeline.ap-northeast-1.amazonaws.com',
+        ),
     ),
     'operations' => array(
         'ActivatePipeline' => array(