+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Next Release
 * Fixed an issue with the Amazon S3 client where an exception was thrown when trying to add a default Content-MD5
   header to a request that uses a non-rewindable stream.
 * Updated the Amazon S3 PostObject class to work with CNAME style buckets.
+* Added the us-west-2, eu-west-1, ap-southeast-2, and ap-northeast-1 regions to the AWS Data Pipeline client.
 
 2.5.2 (2014-01-29)
 ------------------